Extract cordova adapter and create them with factory method
