+++ lib/keycloak.js
@@ -1,7 +1,9 @@
 // @ts-check
 /**
- * @import {Acr, KeycloakAccountOptions, KeycloakAdapter, KeycloakConfig, KeycloakError, KeycloakFlow, KeycloakInitOptions, KeycloakLoginOptions, KeycloakLogoutOptions, KeycloakPkceMethod, KeycloakProfile, KeycloakRegisterOptions, KeycloakResourceAccess, KeycloakResponseMode, KeycloakResponseType, KeycloakRoles, KeycloakTokenParsed, OpenIdProviderMetadata} from "./keycloak.ts"
+ * @import {Acr, KeycloakAccountOptions, KeycloakAdapter, KeycloakAdapterContext, KeycloakConfig, KeycloakError, KeycloakFlow, KeycloakInitOptions, KeycloakLoginOptions, KeycloakLogoutOptions, KeycloakPkceMethod, KeycloakProfile, KeycloakRegisterOptions, KeycloakResourceAccess, KeycloakResponseMode, KeycloakResponseType, KeycloakRoles, KeycloakTokenParsed, OpenIdProviderMetadata} from "./keycloak.ts"
  */
+import { createCordovaAdapter } from "./adapters/cordova.js"
+import { createCordovaNativeAdapter } from "./adapters/cordova-native.js"
 /*
  * Copyright 2016 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
@@ -184,6 +186,9 @@ export default class Keycloak {
 
     if (typeof initOptions.adapter === 'string' && adapters.includes(initOptions.adapter)) {
       this.#adapter = this.#loadAdapter(initOptions.adapter)
+    } else if (typeof initOptions.adapter === 'function') {
+      // Adapter is a factory function - pass it the context
+      this.#adapter = initOptions.adapter(this.#createAdapterContext())
     } else if (typeof initOptions.adapter === 'object') {
       this.#adapter = initOptions.adapter
     } else if ('Cordova' in window || 'cordova' in window) {
@@ -283,6 +288,39 @@ export default class Keycloak {
   }
 
   /**
+   * Creates an adapter context object that provides controlled access to internal methods.
+   * This context is passed to adapter factory functions to enable them to interact with
+   * Keycloak core functionality without exposing these methods as public API.
+   *
+   * @returns {KeycloakAdapterContext} Adapter context with methods for authentication flows
+   */
+  #createAdapterContext() {
+    return {
+      createLoginUrl: /** @type {KeycloakAdapterContext['createLoginUrl']} */ (
+        (options) => this.createLoginUrl(options)
+      ),
+      createLogoutUrl:
+        /** @type {KeycloakAdapterContext['createLogoutUrl']} */ (
+          (options) => this.createLogoutUrl(options)
+        ),
+      createRegisterUrl:
+        /** @type {KeycloakAdapterContext['createRegisterUrl']} */ (
+          (options) => this.createRegisterUrl(options)
+        ),
+      createAccountUrl: () => this.createAccountUrl(),
+      clearToken: () => this.clearToken(),
+      processCallback:
+        /** @type {KeycloakAdapterContext['processCallback']} */ (
+          async (url) => {
+            const callback = this.#parseCallback(url)
+            return await this.#processCallback(callback)
+          }
+        ),
+      redirectUri: this.redirectUri || "http://localhost",
+    }
+  }
+
+  /**
    * @param {"default" | "cordova" | "cordova-native"} type
    * @returns {KeycloakAdapter}
    */
@@ -293,12 +331,12 @@ export default class Keycloak {
 
     if (type === 'cordova') {
       this.#loginIframe.enable = false
-      return this.#loadCordovaAdapter()
+      return createCordovaAdapter(this.#createAdapterContext())
     }
 
     if (type === 'cordova-native') {
       this.#loginIframe.enable = false
-      return this.#loadCordovaNativeAdapter()
+      return createCordovaNativeAdapter(this.#createAdapterContext())
     }
 
     throw new Error('invalid adapter type: ' + type)
@@ -371,273 +409,7 @@ export default class Keycloak {
         return await new Promise(() => {})
       },
 
-      redirectUri
-    }
-  }
-
-  /**
-   * @returns {KeycloakAdapter}
-   */
-  #loadCordovaAdapter () {
-    /**
-     * @param {string} loginUrl
-     * @param {string} target
-     * @param {string} options
-     * @returns {WindowProxy | null}
-     */
-    const cordovaOpenWindowWrapper = (loginUrl, target, options) => {
-      if (window.cordova && window.cordova.InAppBrowser) {
-        // Use inappbrowser for IOS and Android if available
-        return window.cordova.InAppBrowser.open(loginUrl, target, options)
-      } else {
-        return window.open(loginUrl, target, options)
-      }
-    }
-
-    const shallowCloneCordovaOptions = (userOptions) => {
-      if (userOptions && userOptions.cordovaOptions) {
-        return Object.keys(userOptions.cordovaOptions).reduce((options, optionName) => {
-          options[optionName] = userOptions.cordovaOptions[optionName]
-          return options
-        }, {})
-      } else {
-        return {}
-      }
-    }
-
-    const formatCordovaOptions = (cordovaOptions) => {
-      return Object.keys(cordovaOptions).reduce((options, optionName) => {
-        options.push(optionName + '=' + cordovaOptions[optionName])
-        return options
-      }, []).join(',')
-    }
-
-    const createCordovaOptions = (userOptions) => {
-      const cordovaOptions = shallowCloneCordovaOptions(userOptions)
-      cordovaOptions.location = 'no'
-      if (userOptions && userOptions.prompt === 'none') {
-        cordovaOptions.hidden = 'yes'
-      }
-      return formatCordovaOptions(cordovaOptions)
-    }
-
-    const getCordovaRedirectUri = () => {
-      return this.redirectUri || 'http://localhost'
-    }
-
-    return {
-      login: async (options) => {
-        const cordovaOptions = createCordovaOptions(options)
-        const loginUrl = await this.createLoginUrl(options)
-        const ref = cordovaOpenWindowWrapper(loginUrl, '_blank', cordovaOptions)
-        let completed = false
-        let closed = false
-
-        function closeBrowser () {
-          closed = true
-          ref.close()
-        };
-
-        return await new Promise((resolve, reject) => {
-          ref.addEventListener('loadstart', async (event) => {
-            if (event.url.indexOf(getCordovaRedirectUri()) === 0) {
-              const callback = this.#parseCallback(event.url)
-
-              try {
-                await this.#processCallback(callback)
-                resolve()
-              } catch (error) {
-                reject(error)
-              }
-              closeBrowser()
-              completed = true
-            }
-          })
-
-          ref.addEventListener('loaderror', async (event) => {
-            if (!completed) {
-              if (event.url.indexOf(getCordovaRedirectUri()) === 0) {
-                const callback = this.#parseCallback(event.url)
-                try {
-                  await this.#processCallback(callback)
-                  resolve()
-                } catch (error) {
-                  reject(error)
-                }
-                closeBrowser()
-                completed = true
-              } else {
-                reject(new Error('Unable to process login.'))
-                closeBrowser()
-              }
-            }
-          })
-
-          ref.addEventListener('exit', function (event) {
-            if (!closed) {
-              reject(new Error('User closed the login window.'))
-            }
-          })
-        })
-      },
-
-      logout: async (options) => {
-        const logoutUrl = this.createLogoutUrl(options)
-        const ref = cordovaOpenWindowWrapper(logoutUrl, '_blank', 'location=no,hidden=yes,clearcache=yes')
-        let error = false
-
-        ref.addEventListener('loadstart', (event) => {
-          if (event.url.indexOf(getCordovaRedirectUri()) === 0) {
-            ref.close()
-          }
-        })
-
-        ref.addEventListener('loaderror', (event) => {
-          if (event.url.indexOf(getCordovaRedirectUri()) === 0) {
-            ref.close()
-          } else {
-            error = true
-            ref.close()
-          }
-        })
-
-        await new Promise((resolve, reject) => {
-          ref.addEventListener('exit', () => {
-            if (error) {
-              reject(new Error('User closed the login window.'))
-            } else {
-              this.clearToken()
-              resolve()
-            }
-          })
-        })
-      },
-
-      register: async (options) => {
-        const registerUrl = await this.createRegisterUrl()
-        const cordovaOptions = createCordovaOptions(options)
-        const ref = cordovaOpenWindowWrapper(registerUrl, '_blank', cordovaOptions)
-
-        /** @type {Promise<void>} */
-        const promise = new Promise((resolve, reject) => {
-          ref.addEventListener('loadstart', async (event) => {
-            if (event.url.indexOf(getCordovaRedirectUri()) === 0) {
-              ref.close()
-              const oauth = this.#parseCallback(event.url)
-
-              try {
-                await this.#processCallback(oauth)
-                resolve()
-              } catch (error) {
-                reject(error)
-              }
-            }
-          })
-        })
-
-        await promise
-      },
-
-      accountManagement: async () => {
-        const accountUrl = this.createAccountUrl()
-        if (typeof accountUrl !== 'undefined') {
-          const ref = cordovaOpenWindowWrapper(accountUrl, '_blank', 'location=no')
-          ref.addEventListener('loadstart', function (event) {
-            if (event.url.indexOf(getCordovaRedirectUri()) === 0) {
-              ref.close()
-            }
-          })
-        } else {
-          throw new Error('Not supported by the OIDC server')
-        }
-      },
-
-      redirectUri: () => {
-        return getCordovaRedirectUri()
-      }
-    }
-  }
-
-  /**
-   * @returns {KeycloakAdapter}
-   */
-  #loadCordovaNativeAdapter () {
-    /* global universalLinks */
-    return {
-      login: async (options) => {
-        const loginUrl = await this.createLoginUrl(options)
-
-        await new Promise((resolve, reject) => {
-          universalLinks.subscribe('keycloak', async (event) => {
-            universalLinks.unsubscribe('keycloak')
-            window.cordova.plugins.browsertab.close()
-            const oauth = this.#parseCallback(event.url)
-
-            try {
-              await this.#processCallback(oauth)
-              resolve()
-            } catch (error) {
-              reject(error)
-            }
-          })
-
-          window.cordova.plugins.browsertab.openUrl(loginUrl)
-        })
-      },
-
-      logout: async (options) => {
-        const logoutUrl = this.createLogoutUrl(options)
-
-        await new Promise((resolve) => {
-          universalLinks.subscribe('keycloak', () => {
-            universalLinks.unsubscribe('keycloak')
-            window.cordova.plugins.browsertab.close()
-            this.clearToken()
-            resolve()
-          })
-
-          window.cordova.plugins.browsertab.openUrl(logoutUrl)
-        })
-      },
-
-      register: async (options) => {
-        const registerUrl = await this.createRegisterUrl(options)
-
-        await new Promise((resolve, reject) => {
-          universalLinks.subscribe('keycloak', async (event) => {
-            universalLinks.unsubscribe('keycloak')
-            window.cordova.plugins.browsertab.close()
-            const oauth = this.#parseCallback(event.url)
-            try {
-              await this.#processCallback(oauth)
-              resolve()
-            } catch (error) {
-              reject(error)
-            }
-          })
-
-          window.cordova.plugins.browsertab.openUrl(registerUrl)
-        })
-      },
-
-      accountManagement: async () => {
-        const accountUrl = this.createAccountUrl()
-        if (typeof accountUrl !== 'undefined') {
-          window.cordova.plugins.browsertab.openUrl(accountUrl)
-        } else {
-          throw new Error('Not supported by the OIDC server')
-        }
-      },
-
-      redirectUri: (options) => {
-        if (options && options.redirectUri) {
-          return options.redirectUri
-        } else if (this.redirectUri) {
-          return this.redirectUri
-        } else {
-          return 'http://localhost'
-        }
-      }
+      redirectUri,
     }
   }
